pyproject.toml
[project]
name = "custapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioapns==4.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "beautifulsoup4>=4.13.4",
    "cachecontrol==0.14.3",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "cryptography==44.0.2",
    "django==4.1.13",
    "django-filter>=23.5",
    "django-push-notifications==3.2.1",
    "django-unfold>=0.50.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "drf-yasg==1.21.7",
    "firebase-admin==6.8.0",
    "frozenlist==1.6.0",
    "git-filter-repo==2.47.0",
    "google-api-core==2.25.0rc1",
    "google-api-python-client==2.169.0",
    "google-auth==2.40.1",
    "google-auth-httplib2==0.2.0",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.20.2",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "http-ece==1.2.1",
    "httplib2==0.22.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "inflection==0.5.1",
    "msgpack==1.1.0",
    "multidict==6.4.3",
    "mysqlclient==2.2.7",
    "openpyxl>=3.1.5",
    "packaging==24.2",
    "pandas>=2.3.0",
    "pdfkit==1.0.0",
    "pillow==11.1.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "py-vapid==1.9.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pyjwt==2.9.0",
    "pymysql==1.1.1",
    "pyopenssl==25.0.0",
    "pyparsing==3.2.3",
    "pypdf2==3.0.1",
    "python-dotenv==1.1.0",
    "pytz==2025.1",
    "pywebpush==2.0.3",
    "pyyaml==6.0.2",
    "qrcode>=8.2",
    "reportlab==4.3.1",
    "requests==2.32.3",
    "rsa==4.9.1",
    "setuptools==78.1.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "tzdata==2025.1",
    "uritemplate==4.1.1",
    "urllib3==1.26.6",
    "whatsapp-api-client-python>=0.0.49",
    "yarl==1.20.0",
]
